[cp-caps] use unittest to refactor the cp-caps test
